requirements.txt
asgiref==3.4.1
Django==3.2.23
python-dotenv==0.20.0
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.1.1
psycopg2-binary
djangorestframework>=3.14.0
requests>=2.28.0
polyline
geopy
scipy
django-cors-headers
folium
aiohttp
httpx
celery
redis